enhance: example add a font and makefile to be able to build the example
