+++ example/addons.make
@@ -1,0 +1,1 @@
+ofxCenteredTrueTypeFont
