+++ example/src/main.cpp
@@ -1,12 +1,10 @@
 #include "ofMain.h"
 #include "testApp.h"
-#include "ofAppGlutWindow.h"
 
 //========================================================================
 int main( ){
 
-    ofAppGlutWindow window;
-	ofSetupOpenGL(&window, 800, 600, OF_WINDOW);			// <-------- setup the GL context
+	ofSetupOpenGL(800, 600, OF_WINDOW);			// <-------- setup the GL context
 
 	// this kicks off the running of my app
 	// can be OF_WINDOW or OF_FULLSCREEN
